Code cleanup and fix for creating policy under dir mentioned in config
